[fix][all] Потерянные деления на kFull и форма испарения
Перепроверка нашла четыре места, где при переводе в целые числа величины
милли-шкалы перемножены без обратного деления. Все в AnimalSystem:

  meat = kMeatPerSize * size            туша давала 8 000 000 биомассы
                                        вместо 8000 — одна туша кормила
                                        хищника вечно;
  genome.biteSize * size  (дважды)      травоядное съедало куст ЦЕЛИКОМ
                                        каждый тик вместо своей доли,
                                        хищник просил вдесятеро больше
                                        мяса, чем помещается в него;
  kDrinkRate * size                     фляга наполнялась мгновенно.

Правильные соседи стояли рядом и были сделаны верно (energyUpkeep, attack,
kStepEnergy, kSoilDrying) — тем обиднее. Прежняя проверка смотрела на типы
и диапазоны, а не на величины, и прошла полностью: все значения были целые
и в объявленных пределах, просто в тысячу раз не те. Отсюда и числа
поголовья в прошлом отчёте — они мерили ошибку, а не мир.

Испарение воды возвращено из срока в скорость. Срок ("одна тысячная раз в
N тиков") не выражает ничего быстрее одной тысячной за тик, а настраивать
испарение приходится и впятеро быстрее: подобранное вручную значение
0.00149 за тик прежней формой не описывалось вовсе, и верхние четыре
пятых диапазона ползунка стали недостижимы. Теперь это целое "тысячных
глубины за сто тиков", снимаемое разностью двух делений — тем же приёмом,
что и взросление животного: ни накопителя, ни дроби, и выражается любая
скорость.

Старый config.json больше не читается молча. nlohmann, встретив дробное
значение в int-поле, не ругается, а отбрасывает дробную часть:
"grass_coverage": 0.06 превращалось в 0, то есть в мир вообще без травы, а
"mountain_height": 29.53 — в плоскую карту, и узнать об этом было неоткуда.
Ловится сама дробность, а не имена ключей: часть имён уцелела при смене
смысла. Дробному числу в настройках мира взяться неоткуда — значит его
наличие и есть признак файла старой формы.

Восстановлена затёртая конфигурация: при переводе я удалил config.json и
дал серверу создать его заново, то есть заменил подобранные значения
умолчаниями. Каждое возвращено с переводом по своей шкале.

В клиенте черты видов и значения панели наблюдения читаются целыми — не
осталось места, где целое состояние мира снова становится дробью.

+++ clients/desktop/src/NetworkClient.cpp
@@ -381,10 +381,10 @@ void NetworkClient::handleMessage(const std::string& payload) {
                 if (!archetype.is_object()) {
                     continue;
                 }
-                std::vector<std::pair<std::string, float>> traits;
+                std::vector<std::pair<std::string, int>> traits;
                 for (const auto& [name, value] : archetype.items()) {
                     if (name != "species" && value.is_number()) {
-                        traits.emplace_back(name, value.get<float>());
+                        traits.emplace_back(name, value.get<int>());
                     }
                 }
                 working_.plantSpecies.push_back(std::move(traits));
@@ -394,15 +394,15 @@ void NetworkClient::handleMessage(const std::string& payload) {
         // Виды животных и само поголовье — тем же способом, что и трава
         // выше: клиент не знает состава генома и не должен.
         auto readSpecies = [](const nlohmann::json& list) {
-            std::vector<std::vector<std::pair<std::string, float>>> species;
+            std::vector<std::vector<std::pair<std::string, int>>> species;
             for (const auto& archetype : list) {
                 if (!archetype.is_object()) {
                     continue;
                 }
-                std::vector<std::pair<std::string, float>> traits;
+                std::vector<std::pair<std::string, int>> traits;
                 for (const auto& [name, value] : archetype.items()) {
                     if (name != "species" && value.is_number()) {
-                        traits.emplace_back(name, value.get<float>());
+                        traits.emplace_back(name, value.get<int>());
                     }
                 }
                 species.push_back(std::move(traits));
